Refactoring: GridCreatingConditionTriangle::MovePolyLineCommand implementation is separated from gridcreatingconditiontriangle.cpp.
